Bug fixes

Closes #114, #107, #106, #105, and #99

See merge request cse1105/2024-2025/teams/csep-team-67!64

+++ client/src/main/java/client/controllers/FilesCtrl.java
@@ -14,11 +14,14 @@ import javafx.scene.control.*;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import javafx.stage.FileChooser;
+import javafx.util.Duration;
 import lombok.Setter;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.text.DecimalFormat;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 
 
@@ -56,6 +59,17 @@ public class FilesCtrl {
             return null;
         }
 
+        if (!serverUtils.isServerAvailable(currentNote.collection.serverURL)) {
+            String alertText = bundle.getString("noteUpdateError") + "\n" + currentNote.title;
+            dialogStyler.createStyledAlert(
+                    Alert.AlertType.INFORMATION,
+                    bundle.getString("serverCouldNotBeReached.text"),
+                    bundle.getString("serverCouldNotBeReached.text"),
+                    alertText
+            ).showAndWait();
+            return null;
+        }
+
         fileChooser.setTitle(bundle.getString("uploadFile.text"));
         File uploadedFile = fileChooser.showOpenDialog(null);
         if (uploadedFile != null) {
@@ -90,7 +104,7 @@ public class FilesCtrl {
     }
 
     public void showFiles(Note currentNote) {
-        if (currentNote == null)
+        if (currentNote == null || !serverUtils.isServerAvailable(currentNote.collection.serverURL))
             return;
         List<EmbeddedFile> filesInNote = serverUtils.getFilesByNote(currentNote);
         filesView.getChildren().clear();
@@ -153,6 +167,13 @@ public class FilesCtrl {
 
         Label fileName = new Label(file.getFileName());
         fileName.getStyleClass().add("file-view-label");
+        Tooltip infoTooltip = new Tooltip(bundle.getString("saveFile.text") + "\n" +
+                bundle.getString("filename.text") + ": " + file.getFileName() + "\n" +
+                bundle.getString("fileSize.text") + ": " + calculateFileSize(file) + "\n" +
+                bundle.getString("uploadedAt.text") + ": " + file.getUploadedAt().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")));
+        infoTooltip.setShowDelay(Duration.seconds(0.2));
+        fileName.setTooltip(infoTooltip);
+
         fileName.setOnMouseReleased(event -> {
             downloadFile(currentNote, file);
         });
@@ -160,6 +181,9 @@ public class FilesCtrl {
         Button editButton = new Button();
         editButton.getStyleClass().add("icon");
         editButton.getStyleClass().add("file-view-edit-button");
+        Tooltip editTooltip = new Tooltip(bundle.getString("renameFile.text"));
+        editTooltip.setShowDelay(Duration.seconds(0.2));
+        editButton.setTooltip(editTooltip);
         editButton.setOnAction(event -> {
             renameFile(currentNote, file);
         });
@@ -167,6 +191,9 @@ public class FilesCtrl {
         Button deleteButton = new Button();
         deleteButton.getStyleClass().add("icon");
         deleteButton.getStyleClass().add("file-view-delete-button");
+        Tooltip deleteTooltip = new Tooltip(bundle.getString("deleteFile.text"));
+        deleteTooltip.setShowDelay(Duration.seconds(0.2));
+        deleteButton.setTooltip(deleteTooltip);
         deleteButton.setOnAction(event -> {
             dashboardCtrl.getActionHistory().push(
                     new Action(ActionType.DELETE_FILE,currentNote,null,null, file)
@@ -178,6 +205,19 @@ public class FilesCtrl {
         return entry;
     }
 
+    public String calculateFileSize(EmbeddedFile file) {
+        DecimalFormat df = new DecimalFormat("#.##");
+        double bytes = file.getFileContent().length;
+        if (bytes < 1000 /*1kB*/) {
+            return df.format(bytes) + " bytes";
+        }
+        if (bytes < 1000000 /*1MB*/) {
+            return df.format(bytes / 1000) + " kB";
+        }
+        return df.format(bytes / 1000000) + " MB";
+    }
+
+
     public void deleteFile(Note currentNote, EmbeddedFile file) {
         Alert alert = dialogStyler.createStyledAlert(
                 Alert.AlertType.CONFIRMATION,
@@ -187,6 +227,17 @@ public class FilesCtrl {
         );
         Optional<ButtonType> buttonType = alert.showAndWait();
         if (buttonType.isPresent() && buttonType.get().equals(ButtonType.OK)){
+            if (!serverUtils.isServerAvailable(currentNote.collection.serverURL)) {
+                String alertText = bundle.getString("noteUpdateError") + "\n" + currentNote.title;
+                dialogStyler.createStyledAlert(
+                        Alert.AlertType.INFORMATION,
+                        bundle.getString("serverCouldNotBeReached.text"),
+                        bundle.getString("serverCouldNotBeReached.text"),
+                        alertText
+                ).showAndWait();
+                return;
+            }
+
             serverUtils.deleteFile(currentNote, file);
             serverUtils.send("/app/notes/" + currentNote.getId() + "/files/deleteFile", file.getId());
         }
@@ -200,6 +251,16 @@ public class FilesCtrl {
         );
         Optional<String> fileName = dialog.showAndWait();
         if (fileName.isPresent()) {
+            if (fileName.get().isBlank()) {
+                Alert alert = dialogStyler.createStyledAlert(
+                        Alert.AlertType.INFORMATION,
+                        bundle.getString("fileError.text"),
+                        bundle.getString("fileError.text"),
+                        bundle.getString("emptyRenameError")
+                );
+                alert.showAndWait();
+                return;
+            }
             if (!checkFileName(currentNote, fileName.get())) {
                 Alert alert = dialogStyler.createStyledAlert(
                         Alert.AlertType.INFORMATION,
