+++ client/src/main/java/client/controllers/NoteCtrl.java
@@ -7,6 +7,7 @@ import client.utils.Config;
 import client.utils.ServerUtils;
 import com.google.inject.Inject;
 import commons.Collection;
+import commons.EmbeddedFile;
 import commons.Note;
 import javafx.application.Platform;
 import javafx.collections.ObservableList;
@@ -203,6 +204,17 @@ public class NoteCtrl {
             Optional<ButtonType> buttonType = alert.showAndWait();
 
             if (buttonType.isPresent() && buttonType.get().equals(ButtonType.OK)) {
+                if (!server.isServerAvailable(currentNote.collection.serverURL)) {
+                    String alertText = bundle.getString("noteUpdateError") + "\n" + currentNote.title;
+                    dialogStyler.createStyledAlert(
+                            Alert.AlertType.INFORMATION,
+                            bundle.getString("serverCouldNotBeReached.text"),
+                            bundle.getString("serverCouldNotBeReached.text"),
+                            alertText
+                    ).showAndWait();
+                    return;
+                }
+
                 deleteNote(currentNote, collectionNotes, allNotes);
                 noteBody.clear();
                 notificationsCtrl.pushNotification(bundle.getString("deleteSuccess"), false);
@@ -216,6 +228,10 @@ public class NoteCtrl {
                            ObservableList<Note> collectionNotes,
                            ObservableList<Note> allNotes) {
         updatePendingNotes.remove(currentNote);
+        for (EmbeddedFile file : currentNote.getEmbeddedFiles()) {
+            server.deleteFile(currentNote, file);
+            currentNote.getEmbeddedFiles().remove(file);
+        }
         server.send("/app/deleteNote", currentNote);
 
         removeNoteFromClient(currentNote, collectionNotes, allNotes);
@@ -243,7 +259,18 @@ public class NoteCtrl {
     public void saveAllPendingNotes() {
         try {
             for (Note note : updatePendingNotes) {
-                server.updateNote(note);
+                if (!server.isServerAvailable(note.collection.serverURL)) {
+                    String alertText = bundle.getString("noteUpdateError") + "\n" + note.title;
+                    dialogStyler.createStyledAlert(
+                            Alert.AlertType.INFORMATION,
+                            bundle.getString("serverCouldNotBeReached.text"),
+                            bundle.getString("serverCouldNotBeReached.text"),
+                            alertText
+                    ).showAndWait();
+                }
+                else {
+                    server.updateNote(note);
+                }
             }
             updatePendingNotes.clear();
         } catch (Exception e) {
@@ -253,6 +280,17 @@ public class NoteCtrl {
 
     public void onBodyChanged(Note currentNote) {
         if (currentNote != null) {
+            if (!server.isServerAvailable(currentNote.collection.serverURL)) {
+                String alertText = bundle.getString("noteUpdateError") + "\n" + currentNote.title;
+                dialogStyler.createStyledAlert(
+                        Alert.AlertType.INFORMATION,
+                        bundle.getString("serverCouldNotBeReached.text"),
+                        bundle.getString("serverCouldNotBeReached.text"),
+                        alertText
+                ).showAndWait();
+                noteBody.clear();
+                return;
+            }
             String rawText = noteBody.getText();
             currentNote.setBody(rawText);
